Revert "Update to Qt5 and get working on OS X"
This reverts commit aa06d9222552d7a73ac236ad89925d8443d76cdf.

+++ Velocity/transferflagsdialog.cpp
@@ -84,7 +84,7 @@ void TransferFlagsDialog::showRemoveContextMenu(QPoint pos)
        ui->comboBox->addItem(item->text(0));
 
        // add it to the available
-       Flag flag = { item->text(0), static_cast<BYTE>(QtHelpers::ParseHexString(item->text(1))) };
+       Flag flag = { item->text(0),  QtHelpers::ParseHexString(item->text(1))};
        availableFlags.push_back(flag);
 
        delete item;
